remove a couple more mentions of 'ledger'
