+++ src/test/scala/io/iohk/ethereum/blockchain/sync/regular/RegularSyncFixtures.scala
@@ -37,6 +37,7 @@ import scala.collection.mutable
 import scala.concurrent.duration.{DurationInt, FiniteDuration}
 import scala.math.BigInt
 import scala.reflect.ClassTag
+import cats.data.NonEmptyList
 
 // Fixture classes are wrapped in a trait due to problems with making mocks available inside of them
 trait RegularSyncFixtures { self: Matchers with AsyncMockFactory =>
@@ -287,6 +288,45 @@ trait RegularSyncFixtures { self: Matchers with AsyncMockFactory =>
 
     implicit def eqInstanceForPeersClientRequest[T <: Message]: Eq[PeersClient.Request[T]] =
       (x, y) => x.message == y.message && x.peerSelector == y.peerSelector
+
+    class FakeImportBlock extends TestBlockImport {
+      override def importBlock(
+          block: Block
+      )(implicit blockExecutionScheduler: Scheduler): Task[BlockImportResult] = {
+        val result: BlockImportResult = if (didTryToImportBlock(block)) {
+          DuplicateBlock
+        } else {
+          if (
+            importedBlocksSet.isEmpty || bestBlock.isParentOf(block) || importedBlocksSet.exists(_.isParentOf(block))
+          ) {
+            importedBlocksSet.add(block)
+            BlockImportedToTop(List(BlockData(block, Nil, ChainWeight.totalDifficultyOnly(block.header.difficulty))))
+          } else if (block.number > bestBlock.number) {
+            importedBlocksSet.add(block)
+            BlockEnqueued
+          } else {
+            BlockImportFailed("foo")
+          }
+        }
+
+        Task.now(result)
+      }
+    }
+
+    class FakeBranchResolution extends BranchResolution(stub[BlockchainImpl], stub[BlockchainReader]) {
+      override def resolveBranch(headers: NonEmptyList[BlockHeader]): BranchResolutionResult = {
+        val importedHashes = importedBlocksSet.map(_.hash).toSet
+
+        if (
+          importedBlocksSet.isEmpty || (importedHashes.contains(
+            headers.head.parentHash
+          ) && headers.last.number > bestBlock.number)
+        )
+          NewBetterBranch(Nil)
+        else
+          UnknownBranch
+      }
+    }
   }
 
   class OnTopFixture(system: ActorSystem) extends RegularSyncFixture(system) {
